+++ sentry/src/main/java/io/sentry/HubAdapter.java
@@ -86,6 +86,8 @@ public final class HubAdapter implements IHub {
     return Sentry.captureException(throwable, hint, callback);
   }
 
+  @SuppressWarnings("deprecation")
+  @Deprecated
   @Override
   public void captureUserFeedback(@NotNull UserFeedback userFeedback) {
     Sentry.captureUserFeedback(userFeedback);