+++ sentry/src/main/java/io/sentry/NoOpHub.java
@@ -80,6 +80,8 @@ public final class NoOpHub implements IHub {
     return SentryId.EMPTY_ID;
   }
 
+  @SuppressWarnings("deprecation")
+  @Deprecated
   @Override
   public void captureUserFeedback(@NotNull UserFeedback userFeedback) {}