+++ sentry/src/main/java/io/sentry/Scopes.java
@@ -346,6 +346,8 @@ public final class Scopes implements IScopes {
     return sentryId;
   }
 
+  @SuppressWarnings("deprecation")
+  @Deprecated
   @Override
   public void captureUserFeedback(final @NotNull UserFeedback userFeedback) {
     if (!isEnabled()) {